refactor(Participant): Dealer와 Plaer의 부모 클래스 역할을 하는 Participant 클래스 구현
