+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -2,7 +2,7 @@ package blackjack.domain.participant;
 
 import blackjack.domain.card.Card;
 
-public class Dealer extends Player {
+public class Dealer extends Participant {
     private static final int MAX_SUM_FOR_MORE_CARD = 16;
     private static final String name = "딜러";
     private int winCount = 0;
