+++ src/test/java/blackjack/PlayerTest.java
@@ -3,7 +3,9 @@ package blackjack;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardNumber;
 import blackjack.domain.card.CardType;
+import blackjack.domain.participant.Participant;
 import blackjack.domain.participant.Player;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -29,10 +31,18 @@ public class PlayerTest {
     }
 
     @Test
+    @DisplayName("플레이어가 Participant에게 상속받았는지 확인")
+    void extend() {
+        final Participant participant = new Player("bada");
+        participant.receiveCard(new Card(CardNumber.ACE, CardType.HEART));
+        Assertions.assertThat(participant.cardCount()).isEqualTo(1);
+    }
+
+    @Test
     @DisplayName("참가자가 카드를 받았는지 확인")
     void receiveCard() {
         player.receiveCard(new Card(CardNumber.ACE, CardType.CLOVER));
-        assertThat(player.getCardCount()).isEqualTo(1);
+        assertThat(player.cardCount()).isEqualTo(1);
     }
 
     @Test
